Fix CI on Linux and macOS
First push to the public repo went red on two of three platforms. Windows was
green, which is exactly why neither of these had been caught: this was the
first time the code had ever been compiled for the other two.

Linux failed at the apt step, before any Rust ran. The list installed both
`libappindicator3-dev` and `libayatana-appindicator3-dev`. `ubuntu-latest` is
now 24.04, where the former was dropped in favour of the Ayatana fork, so apt
exited non-zero. They are alternative implementations of the same tray API and
installing both was never right. The list now matches Tauri's own prerequisites
page element for element. release.yml had the same list and is fixed too --
it only worked there because it pins ubuntu-22.04.

macOS failed clippy with 10 errors, all of them platform-gated code that had
never been built:

  - The `dpapi` module is unconditional by design so the envelope logic does
    not need cfg(windows) sprinkled through it. Off Windows its stub is
    unreachable, so dead_code/unused_imports are allowed for non-Windows
    targets only -- Windows still fails on genuinely unused items.
  - wsl.rs's pure helpers are compiled everywhere so they stay unit-tested
    everywhere (the UTF-16LE decoder especially), but are only called on
    Windows. Same narrow, per-item allow.
  - `SfError::code()` already returns i32; the cast was a no-op.

Linux still has not gotten past the apt step, so linux.rs remains uncompiled
on its target platform.

+++ src-tauri/src/credentials.rs
@@ -517,6 +517,10 @@ pub fn protection_scheme() -> ProtectionScheme {
 /// `cfg(windows)` in [`wrap_platform_bytes`]/[`unwrap_dpapi`], but keeping
 /// the module unconditional avoids sprinkling `cfg(windows)` through the
 /// envelope logic itself.
+///
+/// Off Windows the stub `imp` is unreachable by construction, so dead-code
+/// lints are allowed there only — Windows still fails on genuinely unused items.
+#[cfg_attr(not(windows), allow(dead_code, unused_imports))]
 mod dpapi {
     #[derive(Debug)]
     pub struct DpapiError(pub String);
@@ -803,7 +807,8 @@ mod macos_keychain {
         const ERR_SEC_ITEM_NOT_FOUND: i32 = -25300;
 
         fn is_not_found(e: &SfError) -> bool {
-            e.code() as i32 == ERR_SEC_ITEM_NOT_FOUND
+            // `code()` is already i32; the cast was a no-op.
+            e.code() == ERR_SEC_ITEM_NOT_FOUND
         }
 
         pub fn get_password(service: &str, account: &str) -> Result<Option<String>, KeychainError> {
